refactor: fix CLI new command technical debt
- Remove unused _slug, _token params from deployProject()
- Replace 500ms timeout hack with server.ready promise
- Create platform shim for waitForKeypress (removes @ts-ignore)
- Add proper error handling in clearProjectFiles()

devCommand now returns { ready, done } promises so callers can
await actual server readiness instead of guessing with timeouts.

Related: veryfront/veryfront-api#118 (empty template support)
